Create InputManager
This commit creates a script and game object called InputManager. I moved the multi-touch disabling from the word bank to this game object. I also moved the code that disables interaction with several game objects while TTS is reading a sentence from the SpeakSentenceButton script to this script.
