+++ Assets/Scenes/Sentence Builder/SentenceBuilder.unity
@@ -2595,7 +2595,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -400.50464, y: 251.99998}
+  m_AnchoredPosition: {x: -400.50464, y: 252.00009}
   m_SizeDelta: {x: 801.01, y: 500}
   m_Pivot: {x: 0, y: 1}
 --- !u!114 &396208986
@@ -7493,7 +7493,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 1969722462}
   m_Direction: 2
   m_Value: 0
-  m_Size: 0.9999973
+  m_Size: 1
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -9888,11 +9888,6 @@ MonoBehaviour:
     type: 3}
   noVoicesImage: {fileID: 114823504137231898, guid: 8753ba4ff7dc7e64990506f776a32d00,
     type: 3}
-  submitSentenceButttonImage: {fileID: 1696695521}
-  currentLearnerSprite: {fileID: 514360513}
-  changeSceneButton: {fileID: 1109885700}
-  wordBankHolder: {fileID: 960626491}
-  mostRecentSavedSentence: {fileID: 461907523}
   sentence: {fileID: 934462495}
   tts: {fileID: 1214538165}
 --- !u!1 &1599891902
@@ -13450,6 +13445,54 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1969722461}
   m_CullTransparentMesh: 0
+--- !u!1 &1982006765
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1982006767}
+  - component: {fileID: 1982006766}
+  m_Layer: 0
+  m_Name: InputManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1982006766
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1982006765}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 54b57f10ce4e4b246bb17e8d952936bb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  submitSentenceButttonImage: {fileID: 1696695521}
+  currentLearnerSprite: {fileID: 514360513}
+  changeSceneButton: {fileID: 1109885700}
+  wordBankHolder: {fileID: 960626491}
+  mostRecentSavedSentence: {fileID: 461907523}
+--- !u!4 &1982006767
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1982006765}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 7
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1999367031
 GameObject:
   m_ObjectHideFlags: 0
